arch(navigation): migrate DeckOptionsDestination to Navigation
Related to 20558
Part of 20737
