+++ AnkiDroid/src/main/java/com/ichi2/anki/navigation/AnkiDroidNavigator.kt
@@ -8,6 +8,7 @@ import com.ichi2.anki.browser.toIntent
 import com.ichi2.anki.common.destinations.BrowserDestination
 import com.ichi2.anki.common.destinations.CardInfoDestination
 import com.ichi2.anki.common.destinations.CsvImporterDestination
+import com.ichi2.anki.common.destinations.DeckOptionsDestination
 import com.ichi2.anki.common.destinations.Destination
 import com.ichi2.anki.common.destinations.Navigator
 import com.ichi2.anki.pages.toIntent
@@ -25,6 +26,7 @@ object AnkiDroidNavigator : Navigator {
             is BrowserDestination -> destination.toIntent(navContext)
             is CardInfoDestination -> destination.toIntent(navContext)
             is CsvImporterDestination -> destination.toIntent(navContext)
+            is DeckOptionsDestination -> destination.toIntent(navContext)
         }
 }
