+++ AnkiDroid/src/main/java/com/ichi2/anki/pages/DeckOptionsDestination.kt
@@ -19,52 +19,30 @@ import android.content.Context
 import android.content.Intent
 import androidx.annotation.CheckResult
 import com.ichi2.anki.CollectionManager.withCol
+import com.ichi2.anki.common.destinations.DeckOptionsDestination
 import com.ichi2.anki.filtered.FilteredDeckOptionsFragment
 import com.ichi2.anki.libanki.DeckId
-import com.ichi2.anki.utils.Destination
 
-/**
- * @param options the list of deck options to present to the user before going to deck options. This
- * will contain the current selected deck([deckId]) plus any other possible deck targets(ex: decks
- * of the current studied card)
- */
-data class DeckOptionsDestination(
-    val deckId: DeckId,
-    val isFiltered: Boolean,
-    val options: List<DeckOptionsEntry> = emptyList(),
-) : Destination {
-    override fun toIntent(context: Context): Intent =
-        if (isFiltered) {
-            FilteredDeckOptionsFragment.getIntent(context, did = deckId)
-        } else {
-            DeckOptions.getIntent(context, deckId)
-        }
-
-    companion object {
-        suspend fun fromDeckId(deckId: DeckId): DeckOptionsDestination =
-            DeckOptionsDestination(
-                deckId = deckId,
-                isFiltered = withCol { decks.isFiltered(deckId) },
-            )
-
-        @CheckResult
-        suspend fun fromCurrentDeck() =
-            withCol {
-                val deckId = decks.getCurrentId()
-                DeckOptionsDestination(
-                    deckId = deckId,
-                    isFiltered = decks.isFiltered(deckId),
-                )
-            }
+/** Builds the [Intent] that opens the deck options screen for this destination. */
+fun DeckOptionsDestination.toIntent(context: Context): Intent =
+    if (isFiltered) {
+        FilteredDeckOptionsFragment.getIntent(context, did = deckId)
+    } else {
+        DeckOptions.getIntent(context, deckId)
     }
-}
 
-/**
- * Information about a deck that appears in the list of possible deck targets when deck options are
- * requested from the study screen.
- */
-data class DeckOptionsEntry(
-    val deckId: DeckId,
-    val name: String?,
-    val isFiltered: Boolean,
-)
+suspend fun DeckOptionsDestination.Companion.fromDeckId(deckId: DeckId): DeckOptionsDestination =
+    DeckOptionsDestination(
+        deckId = deckId,
+        isFiltered = withCol { decks.isFiltered(deckId) },
+    )
+
+@CheckResult
+suspend fun DeckOptionsDestination.Companion.fromCurrentDeck(): DeckOptionsDestination =
+    withCol {
+        val deckId = decks.getCurrentId()
+        DeckOptionsDestination(
+            deckId = deckId,
+            isFiltered = decks.isFiltered(deckId),
+        )
+    }
