+++ AnkiDroid/src/main/java/com/ichi2/widget/cardanalysis/CardAnalysisWidget.kt
@@ -29,10 +29,13 @@ import com.ichi2.anki.R
 import com.ichi2.anki.analytics.UsageAnalytics
 import com.ichi2.anki.common.coroutines.applicationScope
 import com.ichi2.anki.common.crashreporting.CrashReportService
+import com.ichi2.anki.common.destinations.DeckOptionsDestination
+import com.ichi2.anki.common.destinations.DeferredNavigation
+import com.ichi2.anki.common.destinations.toIntent
 import com.ichi2.anki.isCollectionEmpty
 import com.ichi2.anki.libanki.DeckId
 import com.ichi2.anki.libanki.Decks.Companion.NOT_FOUND_DECK_ID
-import com.ichi2.anki.pages.DeckOptionsDestination
+import com.ichi2.anki.pages.fromDeckId
 import com.ichi2.widget.ACTION_UPDATE_WIDGET
 import com.ichi2.widget.AnalyticsWidgetProvider
 import com.ichi2.widget.AppWidgetId
@@ -196,7 +199,7 @@ class CardAnalysisWidget : AnalyticsWidgetProvider() {
                 if (!isEmptyDeck) {
                     intentToReviewDeckFromShortcuts(context, deckData.deckId)
                 } else {
-                    DeckOptionsDestination.fromDeckId(deckData.deckId).toIntent(context)
+                    with(DeferredNavigation) { DeckOptionsDestination.fromDeckId(deckData.deckId).toIntent() }
                 }
             val pendingIntent =
                 PendingIntent.getActivity(
